+++ tests/db/test_table_model.py
@@ -104,3 +104,21 @@ def test_model_filters_records(sakila_connection, filter):
 def test_model_properly_escapes_dash_in_table_names(db_dash_in_name_connection):
     model = TableModel(db_dash_in_name_connection, 'table-with-dash')
     assert model.last_error is None
+
+def test_model_properly_escapes_alias(sakila_connection):
+    long_str = "x" * 1024
+    create_table_sql = '''
+CREATE TABLE IF NOT EXISTS `table_with_space_in_column_names` (
+counter INT NULL,
+`column varchar` VARCHAR (1024) NULL DEFAULT "{0}",
+`column text` TEXT NULL DEFAULT ("{0}")
+)
+    '''.format(long_str)
+
+    sakila_connection.query(create_table_sql)
+    sakila_connection.query('REPLACE INTO table_with_space_in_column_names (counter) VALUES (1)')
+
+    model = TableModel(sakila_connection, 'table_with_space_in_column_names')
+    assert model.last_error is None
+
+