+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Smart autocomplete for common statements
+- Add "Loading..." status when switching views
+
+### Fixed
+- Fix a bug where a column containing spaces in its name breaks the table screen view
 
 ## [1.3.0]